requirements.txt
selenium==4.15.2
perplexity-ai
webdriver-manager==4.0.1
requests==2.31.0
urllib3==2.0.7